Updated powershell deployment and ADO pipeline (#50)
* Update Development.json

Update development instance manifest to latest versions

* Instance-level infrastructure now includes a keyvault

This stores the password for the default tenant administrator service principle that gets created for use with the Marain CLI.

Only the deployment user gets access to it.

* Added some Az authentication handling

This should hopefully cause a 'fail early' in the event that the script is not already authenticated.

Also adds probing of the graph api to try and force getting a suitable token, if available.

* Updated main deployment script to support the instance-level keyvault

Ensures the deployment user is granted access to the keyvault
Makes the keyvault's details available via the context

* Added 'tenant admin' principal details to context

The tenancy deployment now creates a default 'tenant administrator' principal for use with the CLI.

This additional context makes the required configuration available to other deployments so they can register their service tenants (e.g. operations & workflow)

* Bolstered handling of the 'no AAD access' scenario

Uses the same approach but adds a 'fail early' guard rail and makes the mode available via the context

* Added root exception handling to improve debugging experience

Adds a mother-of-all try/catch in the main deployment script to ensure we can provide additional context for anyone troubleshooting the powershell

* Added path to marain cli

This is likely temporary until we have an alternative method of distributing the cli.

You will need to manually copy the cli into a 'Solutions/Marain.Instance.Deployment/tools/maraincli' folder

* Added early fail guard rails for azure-cli

Ensures that the azure-cli is available and logged-in

* Added early fail guard rails for azure-cli (again)

* Correct bad type. Update manifest.

* Updated basic teardown script

This should make it quicker and 'safer' to completely teardown an environment

* Obtain marain cli as a dotnet global tool

* Initial ADO release pipeline

* debug the environment

* Added explicit artifact download

* Testing variables and new service connection

* Test calling the powershell deployment script

* Fix working path

* Pipeline debugging

* Pipeline debugging

* Pipeline debugging

* Pipeline debugging

* pipeline debugging

* pipeline debugging

* Testing approach for handling azure-cli login

* pipeline debugging

* Tweak azure-cli login logic to work when running in ADO

* Change how the deployment script is launched

AzurePowershell task doesn't let you customise the environment

* Added null check to handle when app service AppId not passed

* Added support for the 'omit' option in manifests

* Updated releases in Dev manifest

* Testing new ADO service connection

* pipeline SPN debugging

* Updated AAD graph access check

Service principal tokens appear unassociated to the tenant

* Fix pipeline typo

* Added script to automate ADO setup

A new SPN is created, assigned the required permissions and then registered as a Azure DevOps Service Connection

* Removed some debug logging

* Bump Marain.Tenancy version

* Fixed invalid variable reference

* Added endjin-dev stage

* Added retry logic when assigning app permissions

This aims to workaround an apparent timing issue when this runs too soon after the app service deployment.

* Updated retry logic

Looks like we might need to wait longer

* Testing pipeline step template

* Fix parameter references

* pipeline debugging

* Testing AzureAD standard preview module in ADO

* Fixed typo

* Testing AAD permission change

Reverting to previous role assignment implementation

* Test calling AzureAD module out-of-process

To avoid the assembly version conflict between the AzureAd and Az modules

* Add guard rails for AzureAD module installation

* Adding a better guard rail

* Fixed check for existing role assignment

* pipeline debugging

* pipeline debugging

* pipeline debugging

* pipeline debugging

* pipeline debugging

* pipeline debugging

* Ignoring app role idempotency for the moment

The dynamic powershell works fine locally and running in a linux container, but the 'Get-AzureADServiceAppRoleAssignment' always returns null when running in ADO.

Ignoring errors and logging any as warnings for the moment

* debugging pipleline

* Bumped workflow version

* Fix workflow version typo

* Bumped workflow version

* Bumped workflow version

* teardown script tweaks

* Moved ADO service connection script to another repo

* Allow the marain prefix to be more easily overridden

* Fixed typo in pipeline step template

* Added retry logic around ARM deployments

We regularly see spurious ARM errors during deployments

* Ensure unique ARM deploy names on retries

* Bumped 'Development' package versions to latest previews

* Updates to ensure unique deployment names on retry

* Tweaks to retry logic

* Bump component version numbers

* Renamed pipeline

* Bumped versions after merges

* Remove our real deployments from the pipeline

* Rename github ADO endpoint

* Fix pipeline typo

* Reference new ADO service connection

* Fix tagging behaviour

* Fix template references

* Fix pipeline step vmImage

* Investigating extended deployment times

* Updated ARM deployment logging method

It seems that `Write-Warning` messages are not logged the same way as `Write-Host` (deferred?).  This seems to have been causing instances of ARM deployment retries to be effectively invisible in the ADO logs and leading to what appeared to be extended deployments.

* More troubleshooting of delays during ARM deployments in ADO

* More extended deployment troubleshooting

Testing behaviour with a windows agent
Only tag after a successful test deployment

* Switch back to linux agent

Disabled progress messaging and removed `preferredAzurePowerShellVersion` references

* AzurePowerShell version experiments

* Re-enable verbose logging for ARM deployments

* Updated pipeline to cleardown the test deployment

Co-authored-by: Jonathan George <jon@jonathangeorge.co.uk>